requirements.txt
PyYAML >= 3.1.12
aiohttp >= 1.3.3